Extract files from serialize util tests
