+++ packages/typespec-ts/test/modularUnit/serializeUtil/modular_special_union_deserialization/should_generate_deserialize_util_even_if_circular_in_model_properties_but_no_other_special_variants/serializeUtil.ts
@@ -1,0 +1,35 @@
+import { PetUnion, Gold, DogUnion } from "../models/models.js";
+import { PetOutput, GoldOutput, DogOutput } from "../rest/index.js";
+
+/** deserialize function for PetOutput */
+export function deserializePetUnion(obj: PetOutput): PetUnion {
+  switch (obj.kind) {
+    case "dog":
+      return deserializeDogUnion(obj as DogUnion);
+    default:
+      return obj;
+  }
+}
+
+/** deserialize function for Gold */
+function deserializeGold(obj: GoldOutput): Gold {
+  return {
+    kind: obj["kind"],
+    type: obj["type"],
+    bark: obj["bark"],
+    name: obj["name"],
+    weight: obj["weight"],
+    friends: obj["friends"].map((p) => deserializePetUnion(p)),
+    birthDay: new Date(obj["birthDay"])
+  };
+}
+
+/** deserialize function for DogOutput */
+export function deserializeDogUnion(obj: DogOutput): DogUnion {
+  switch (obj.type) {
+    case "gold":
+      return deserializeGold(obj as Gold);
+    default:
+      return obj;
+  }
+}
