+++ packages/typespec-ts/test/modularUnit/serializeUtil/modular_special_union_deserialization/should_not_generate_deserialize_util_but_generate_as_any_if_there's_a_special_union_variant_of_model_with_bytes_property_without_discriminator/test.tsp
@@ -1,0 +1,21 @@
+model WidgetData0 {
+  fooProp: string;
+}
+
+model WidgetData1 {
+  data: bytes;
+}
+
+model Widget {
+  @key id: string;
+  weight: int32;
+  color: "red" | "blue";
+}
+
+model Widget1 extends Widget {
+  data: WidgetData0 | WidgetData1;
+}
+
+interface WidgetService {
+  @get @route("customGet1") customGet1(): Widget1;
+}
