+++ packages/typespec-ts/test/modularUnit/serializeUtil/modular_special_union_deserialization/should_not_generate_deserialize_util_even_if_circular_in_model_properties_but_no_other_special_variants/test.tsp
@@ -1,0 +1,23 @@
+@discriminator("kind")
+model Pet {
+  kind: string;
+  name: string;
+  weight?: float32;
+}
+model Cat extends Pet {
+  kind: "cat";
+  meow: int32;
+}
+@discriminator("type")
+model Dog extends Pet {
+  kind: "dog";
+  type: string;
+  bark: string;
+}
+model Gold extends Dog {
+  type: "gold";
+  friends: Pet[];
+}
+op read(): {
+  @body body: Pet;
+};
