+++ packages/typespec-ts/test/modularUnit/serializeUtil/modular_special_union_deserialization/shouldn't_generate_deserialize_util_or_as_any_if_there's_no_special_union_variant_with_discriminator/test.tsp
@@ -1,0 +1,28 @@
+@discriminator("kind")
+union WidgetData {
+    kind0: WidgetData0,
+    kind1: WidgetData1,
+}
+model WidgetData0 {
+    kind: "kind0";
+    fooProp: string;
+}
+
+model WidgetData1 {
+    kind: "kind1";
+    barProp: string;
+}
+
+model Widget {
+    @key id: string;
+    weight: int32;
+    color: "red" | "blue";
+}
+
+model Widget1 extends Widget {
+    data: WidgetData;
+}
+
+interface WidgetService {
+    @get @route("customGet1") customGet1(): Widget1;
+}
