+++ packages/typespec-ts/test/modularUnit/serializeUtil/modular_special_union_deserialization_should_generate_deserialize_util_even_if_circular_in_model_properties_but_no_other_special_variants/test.tsp
@@ -1,0 +1,30 @@
+@discriminator("kind")
+    union WidgetData {
+      kind0: WidgetData0;
+      kind1: WidgetData1;
+    }
+
+    model WidgetData0 {
+      kind: "kind0";
+      fooProp: string;
+    }
+    
+    model WidgetData1 {
+      kind: "kind1";
+      start: utcDateTime;
+      end?: utcDateTime;
+    }
+    
+    model Widget {
+      @key id: string;
+      weight: int32;
+      color: "red" | "blue";
+    }
+    
+    model Widget1 extends Widget {
+      data: WidgetData
+    }
+
+    interface WidgetService {
+      @get @route("customGet1") customGet1(@body body: Widget1): void;
+    }
