+++ packages/typespec-ts/test/modularUnit/serializeUtil/modular_special_union_serialization/should_generate_serialize_util_if_there're_special_discriminated_union_variants_of_model_with_bytes_and_datetime_property/serializeUtil.ts
@@ -1,0 +1,32 @@
+import { uint8ArrayToString } from "@azure/core-util";
+import {
+  WidgetData0 as WidgetData0Rest,
+  WidgetData1 as WidgetData1Rest,
+  WidgetData as WidgetDataRest
+} from "../rest/index.js";
+import { WidgetData0, WidgetData1, WidgetData } from "../models/models.js";
+
+/** serialize function for WidgetData0 */
+function serializeWidgetData0(obj: WidgetData0): WidgetData0Rest {
+  return {
+    kind: obj["kind"],
+    fooProp: uint8ArrayToString(obj["fooProp"], "base64")
+  };
+}
+
+/** serialize function for WidgetData1 */
+function serializeWidgetData1(obj: WidgetData1): WidgetData1Rest {
+  return { kind: obj["kind"], data: obj["data"].toISOString() };
+}
+
+/** serialize function for WidgetData */
+export function serializeWidgetData(obj: WidgetData): WidgetDataRest {
+  switch (obj.kind) {
+    case "kind0":
+      return serializeWidgetData0(obj);
+    case "kind1":
+      return serializeWidgetData1(obj);
+    default:
+      return obj;
+  }
+}
