+++ packages/typespec-ts/test/modularUnit/serializeUtil/modular_special_union_serialization/should_generate_serialize_util_if_there're_special_discriminated_union_variants_of_model_with_bytes_and_datetime_property/test.tsp
@@ -1,0 +1,29 @@
+@discriminator("kind")
+union WidgetData {
+    kind0: WidgetData0,
+    kind1: WidgetData1,
+}
+
+model WidgetData0 {
+    kind: "kind0";
+    fooProp: bytes;
+}
+
+model WidgetData1 {
+    kind: "kind1";
+    data: utcDateTime;
+}
+
+model Widget {
+    @key id: string;
+    weight: int32;
+    color: "red" | "blue";
+}
+
+model Widget1 extends Widget {
+    data: WidgetData;
+}
+
+interface WidgetService {
+    @get @route("customGet1") customGet1(@body body: Widget1): void;
+}
