+++ packages/typespec-ts/test/modularUnit/serializeUtil/modular_special_union_serialization/should_generate_serialize_util_if_there's_a_special_discriminated_union_variant_of_model_with_datatime_property/serializeUtil.ts
@@ -1,0 +1,24 @@
+import {
+  WidgetData1 as WidgetData1Rest,
+  WidgetData as WidgetDataRest
+} from "../rest/index.js";
+import { WidgetData1, WidgetData } from "../models/models.js";
+
+/** serialize function for WidgetData1 */
+function serializeWidgetData1(obj: WidgetData1): WidgetData1Rest {
+  return {
+    kind: obj["kind"],
+    start: obj["start"].toISOString(),
+    end: obj["end"]?.toISOString()
+  };
+}
+
+/** serialize function for WidgetData */
+export function serializeWidgetData(obj: WidgetData): WidgetDataRest {
+  switch (obj.kind) {
+    case "kind1":
+      return serializeWidgetData1(obj);
+    default:
+      return obj;
+  }
+}
