+++ packages/typespec-ts/test/modularUnit/serializeUtil/modular_special_union_serialization/should_not_generate_serialize_util_but_generate_as_any_if_there's_a_special_union_variant_of_model_with_datatime_property_without_discriminator/test.tsp
@@ -1,0 +1,22 @@
+model WidgetData0 {
+  fooProp: string;
+}
+
+model WidgetData1 {
+  start: utcDateTime;
+  end?: utcDateTime;
+}
+
+model Widget {
+  @key id: string;
+  weight: int32;
+  color: "red" | "blue";
+}
+
+model Widget1 extends Widget {
+  data: WidgetData0 | WidgetData1;
+}
+
+interface WidgetService {
+  @get @route("customGet1") customGet1(@body body: Widget1): void;
+}
